# docs/swagger.yaml
definitions:
...
  core_dtos.TwCreateScheduleRequest:
    properties:
      board_column_id:
        description: Nullable field
        type: integer
      description:
        description: Nullable field
        type: string
      title:
        description: Nullable field
        type: string
      workspace_id:
        description: Nullable field
        type: integer
      workspace_user_id:
        type: integer
    type: object
  core_dtos.TwCreateShecduleResponse:
    properties:
      board_column_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      position:
        type: integer
      title:
        type: string
      workspace_id:
        type: integer
    type: object
...
  core_dtos.TwScheduleResponse:
    properties:
      all_day:
        type: boolean
      board_column_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      end_time:
        type: string
      extra_data:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      location:
        type: string
      position:
        type: integer
      priority:
        type: string
      recurrence_pattern:
        type: string
      start_time:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      video_transcript:
        type: string
      visibility:
        type: string
      workspace_id:
        type: integer
    type: object
  core_dtos.TwUpdateSchedulePosition:
    properties:
      board_column_id:
        type: integer
      position:
        type: integer
    type: object
  core_dtos.TwUpdateScheduleRequest:
    properties:
      all_day:
        type: boolean
      description:
        type: string
      end_time:
        type: string
      extra_data:
        type: string
      location:
        type: string
      priority:
        description: Position          *int       `json:"position,omitempty"`
        type: string
      recurrence_pattern:
        type: string
      start_time:
        type: string
      status:
        type: string
      title:
        description: BoardColumnID     *int       `json:"board_column_id,omitempty"`
        type: string
      video_transcript:
        type: string
      visibility:
        type: string
      workspace_id:
        type: integer
    type: object
  core_dtos.TwUpdateScheduleResponse:
    properties:
      all_day:
        type: boolean
      board_column_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      end_time:
        type: string
      extra_data:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      location:
        type: string
      position:
        type: integer
      priority:
        type: string
      recurrence_pattern:
        type: string
      start_time:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      video_transcript:
        type: string
      visibility:
        type: string
      workspace_id:
        type: integer
    type: object
...
  document_dtos.TwDocumentResponse:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      download_url:
        type: string
      email:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      profile_picture:
        type: string
      role:
        type: string
      schedule_id:
        type: integer
      status_workspace_user:
        type: string
      updated_at:
        type: string
      uploaded_at:
        type: string
      user_id:
        type: integer
      workspace_user_id:
        type: integer
    type: object
  fiber.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: Timewise API
  title: timewise-api
  version: "1.0"
paths:
...
  /api/v1/document/delete:
    delete:
      consumes:
      - application/json
      description: Delete file from Google Cloud Storage
      parameters:
      - description: Schedule ID associated with the file
        in: query
        name: scheduleId
        required: true
        type: string
      - description: Name of the file to delete
        in: query
        name: fileName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: File deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request - Missing or invalid parameters
          schema:
            type: string
        "500":
          description: Internal Server Error - Something went wrong during file deletion
          schema:
            type: string
      security:
      - bearerToken: []
      summary: Delete file
      tags:
      - document
  /api/v1/document/download/{documentId}:
    get:
      description: Download a document from Google Cloud Storage
      parameters:
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File downloaded successfully
          schema:
            type: file
        "404":
          description: Document not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - bearerToken: []
      summary: Download document
      tags:
      - document
  /api/v1/document/schedule/{schedule_id}:
    get:
      consumes:
      - application/json
      description: Get documents by schedule
      parameters:
      - description: Schedule ID
        in: path
        name: schedule_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/document_dtos.TwDocumentResponse'
            type: array
      summary: Get documents by schedule
      tags:
      - document
  /api/v1/document/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload file with multipart/form-data and upload to Google Cloud
        Storage
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Schedule ID associated with the file
        in: formData
        name: scheduleId
        required: true
        type: string
      - description: Workspace user ID who uploads the file
        in: formData
        name: wspUserId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: File uploaded successfully
          schema:
            type: string
        "400":
          description: Bad Request - Missing or invalid parameters
          schema:
            type: string
        "500":
          description: Internal Server Error - Something went wrong during file upload
          schema:
            type: string
      security:
      - bearerToken: []
      summary: Upload file
      tags:
      - document
...
  /api/v1/schedule/schedule:
    get:
      consumes:
      - application/json
      description: Retrieve a list of schedules based on specified filter parameters.
      parameters:
      - description: Filter by Workspace ID
        in: query
        name: workspace_id
        type: string
      - description: Filter by Board Column ID
        in: query
        name: board_column_id
        type: integer
      - description: Filter by Title (searches with LIKE)
        in: query
        name: title
        type: string
      - description: Filter by Start Time (ISO8601 format)
        in: query
        name: start_time
        type: string
      - description: Filter by End Time (ISO8601 format)
        in: query
        name: end_time
        type: string
      - description: Filter by Location (searches with LIKE)
        in: query
        name: location
        type: string
      - description: Filter by User ID of the creator
        in: query
        name: created_by
        type: integer
      - description: Filter by Status
        in: query
        name: status
        type: string
      - description: Filter by Deleted Schedules
        in: query
        name: is_deleted
        type: boolean
      - description: Filter by User ID assigned to the schedule
        in: query
        name: assigned_to
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of filtered schedules
          schema:
            items:
              $ref: '#/definitions/core_dtos.TwScheduleResponse'
            type: array
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/fiber.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fiber.Error'
      security:
      - bearerToken: []
      summary: Get schedules by filter
      tags:
      - schedule
...
  /api/v1/schedules:
    post:
      consumes:
      - application/json
      description: Create a new schedule
      parameters:
      - description: Schedule
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/core_dtos.TwCreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/core_dtos.TwCreateShecduleResponse'
      summary: Create a new schedule
      tags:
      - schedule
  /api/v1/schedules/{schedule_id}:
    delete:
      consumes:
      - application/json
      description: Delete a schedule
      parameters:
      - description: Schedule ID
        in: path
        name: schedule_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete a schedule
      tags:
      - schedule
    get:
      consumes:
      - application/json
      description: Get schedule by ID
      parameters:
      - description: Schedule ID
        in: path
        name: schedule_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core_dtos.TwScheduleResponse'
      summary: Get schedule by ID
      tags:
      - schedule
    put:
      consumes:
      - application/json
      description: Update an existing schedule
      parameters:
      - description: Schedule ID
        in: path
        name: schedule_id
        required: true
        type: integer
      - description: Schedule
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/core_dtos.TwUpdateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core_dtos.TwUpdateScheduleResponse'
      summary: Update an existing schedule
      tags:
      - schedule
  /api/v1/schedules/position/{schedule_id}:
    put:
      consumes:
      - application/json
      description: Update an existing schedule position
      parameters:
      - description: Schedule ID
        in: path
        name: schedule_id
        required: true
        type: integer
      - description: Schedule
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/core_dtos.TwUpdateSchedulePosition'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core_dtos.TwUpdateScheduleResponse'
      summary: Update an existing schedule position
      tags:
      - schedule
...
securityDefinitions:
  bearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"